style: home page and documentation page styled
