+++ src/layout/NavBar/index.css
@@ -1,0 +1,40 @@
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.8em 2.5em;
+}
+
+.all-nav-link{
+    display: flex;
+    gap: 1em;
+}
+
+.nav-link{
+    font-size: 0.8em;
+    font-weight: 600;
+    text-decoration: none;
+}
+
+.logo {
+    display: flex;
+    gap: 0.2em;
+    align-items: center;
+    color: var(--black);
+    text-decoration: none;
+    font-weight: 700;
+}
+
+.logo > img{
+    display: flex;
+    width: 1em;
+    padding-top: 0.1em;
+}
+
+.logo:hover{
+    color: var(--primary);
+}
+
+.logo:active{
+    color: var(--primary-dark);
+}
