--- src/pages/Home.jsx
@@ -1,7 +1,0 @@
-
-
-const Home = () =>{
-    return <div>Home</div>
-}
-
-export default Home;
