+++ src/pages/Home/index.css
@@ -1,0 +1,42 @@
+.home-main {
+  display: flex;
+  align-items: center;
+  gap: 1em;
+  padding: 2em 3em;
+}
+
+.hero-content {
+  width: 50%;
+}
+
+.hero-image {
+  max-width: 50%;
+  margin: 1em 0;
+}
+
+.hero-text > span {
+  color: var(--primary);
+  white-space: nowrap;
+}
+
+.hero-para {
+  font-size: 0.8em;
+  padding: 1em 0;
+}
+
+.home-btn {
+  display: inline-block;
+  padding: 0.5em 1.5em;
+  font-weight: 600;
+  font-size: 0.8em;
+  text-decoration: none;
+  background-color: var(--primary);
+  color: var(--white);
+  border-radius: 2em;
+  box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px,
+    rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
+}
+
+.home-btn:hover {
+  background-color: var(--primary-dark);
+}
